Turn the `gen_brillig_for` method into a function
